Fix trailing comment visitation and add coverage for trailing skip comments on function definitions
